Add files via upload
Business Problem

In recent years, City Hotel and Resort Hotel have seen high cancellation rates. Each hotel is now dealing with a number of issues as a result, including fewer revenues and less than ideal hotel room use. Consequently, lowering cancellation rates is both hotels' primary goal in order to increase their efficiency in generating revenue, and for us to offer thorough business advice to address this problem. The analysis of hotel booking cancellations as well as other factors that have no bearing on their business and yearly revenue generation are the main topics of this report.

Suggestion:

cancellation rates rise as the price does. in order to prevent cancellation of reservations, hotels could work on their pricing strategies and try to lower the rates for specific hotels based on locations. they can also provide some discounds to the consumers.

As the ratio of cancellation and not cancellation of the resort hotel is higher in the resort hotel then the city hotel. So, the hotels should provide a reasonable discount on room prices on weekends or on holidays.

In the month of january, the hotels can start campaigns or marketing with reasonable amount to increase their revenue as the cancellation is the highest in this month.

They can also increase the quality of their hotels and their services mainly in Portugal to reduce the cancellation rate.